Fix missing spaces in PrintDNSResults
You are missing spaces in the PrintDNSResults function. When it runs, the validator will output 'inhosts', 'incache', 'inDNS, and 'inOpen' with no space after the word 'in'. You also need a space in the ValidateDNS function for the same issue, currently there is no space between the word 'for' and the DNS record.